+++ packages/template-retail-react-app/app/contexts/index.js
@@ -9,13 +9,7 @@ import React, {useEffect, useState} from 'react'
 import PropTypes from 'prop-types'
 import {useCommerceAPI} from '../commerce-api/contexts'
 import {CAT_MENU_STALE_TIME} from '../constants'
-import {
-    useCustomer,
-    useCustomerId,
-    useCustomerProductLists,
-    useCustomerType
-} from 'commerce-sdk-react-preview'
-import {CustomerProductListsProvider, useProductLists} from './customer-product-lists'
+import {useCustomer, useCustomerId, useCustomerType} from 'commerce-sdk-react-preview'
 
 /**
  * This is the global state for categories, we use this for navigation and for
@@ -224,7 +218,3 @@ export const CustomerProvider = ({children}) => {
 CustomerProvider.propTypes = {
     children: PropTypes.node.isRequired
 }
-
-export const CustomerProductListsContext = React.createContext()
-
-export {CustomerProductListsProvider, useProductLists}
